// types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IAboutFields {
  /** Participant */
  participant: IParticipant[];

  /** About */
  about: string;

  /** Instagram */
  instagram: string;
}

/** About section */

export interface IAbout extends Entry<IAboutFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'about';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IAlbumFields {
  /** Name */
  name: string;

  /** Year */
  year: string;

  /** Link */
  link: string;

  /** Cover */
  cover: Asset;
}

export interface IAlbum extends Entry<IAlbumFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'album';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IConcertFields {
  /** Place */
  place: string;

  /** City */
  city: string;

  /** Date */
  date: string;

  /** Location */
  location?: string | undefined;
}

export interface IConcert extends Entry<IConcertFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'concert';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IGalleryFields {
  /** Name */
  name: string;

  /** Photo */
  photo: Asset[];
}

export interface IGallery extends Entry<IGalleryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'gallery';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHeroFields {
  /** Background */
  bg: Asset;

  /** Logo */
  logo: Asset;
}

/** Hero section */

export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'hero';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IParticipantFields {
  /** Name */
  name: string;

  /** Instagram */
  instagram: string;

  /** Photo */
  photo: Asset;
}

/** Group member */

export interface IParticipant extends Entry<IParticipantFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'participant';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'about' | 'album' | 'concert' | 'gallery' | 'hero' | 'participant';

export type LOCALE_CODE = 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';